just use String(read(f)), not String(Mmap.mmap(f))
If you are just going to read the whole file and copy it into a string, there is no advantage to `mmap` as far as I can tell.  On my machine, for a 1GB file, the `mmap` version is almost 2x slower, and for a few MB file it is 30% slower.

This also fixes a bug where a code path was using the nonexistent `String(pointer, len)` function.   The intended function is apparently `unsafe_string`, but this is not gc-safe unless you do `GC.@preserve` on the object the pointer was taken from, and is not necessary anyway since you can just do `String(mmap_array)`.   This bug was [pointed out on discourse](https://discourse.julialang.org/t/csvfiles-read-file-with-non-standard-extension/18096).